pyproject.toml
[tool.poetry]
name = "iva-hr-oco-demo-bot"
version = "1"
description = "iva-hr-oco-demo-bot"
authors = ["widegenesis <widegenesis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
ormar = {extras = ["sqlite"], version = "^0.11.0"}
botbuilder-core = "^4.14.2"
botbuilder-schema = "^4.14.2"
botbuilder-dialogs = "^4.14.2"
aiohttp = "^3.8.1"
botbuilder-azure = "^4.14.2"
azure-cognitiveservices-speech = "^1.22.0"
aiofiles = "^0.8.0"
pydub = "^0.25.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
